refactor(cli-agent): refactor agent commands in a separate module to reduce code conflicts
